GoFundMe/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-embed-video==1.4.0
django-widget-tweaks==1.4.8
idna==2.10
Pillow==8.1.0
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
